always track the _raw_ instance
this may yet not be enough for callback-oriented methods such
as PointTree.visitDocValues(). If we continue to hit errors,
we'll simply have to inc/decRef at beginning/end of such methods.

But for now we know we have a problem in the way we're wrapping
the raw readers, since the wrapped readers are very explicitly
unaware of being wrapped, so if we track the wrapper refs, they
become GC'able well before the raw instances that do the actual
work.

In any case we'll proceed iteratively here so that we have a
better sense of where the problem is, and where it's been fixed.